refactor : 도메인 패키지 변경에 따른 테스트 코드 패키지 변경
