+++ src/test/java/subway/domain/line/LineTest.java
@@ -1,8 +1,7 @@
-package subway.domain;
+package subway.domain.line;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.domain.line.Line;
 import subway.domain.section.Section;
 import subway.domain.station.Station;
 import subway.domain.station.Stations;
