+++ src/test/java/subway/domain/station/StationsTest.java
@@ -1,9 +1,7 @@
-package subway.domain;
+package subway.domain.station;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.domain.station.Station;
-import subway.domain.station.Stations;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
